Update optimizer.py
see changes to jobs/process/BaseSDTrainProcess.py which is needed in order for this to work.  It simply edits a redundant lower() function so that user input case can be preserved for the dynamic import.

this edit makes it so that ai toolkit can now accept user-defined optimizers.  

It will still use the same optimizers you had hard-coded so that it doesn't break current configs.